@@ -9,15 +9,15 @@
     "@testing-library/jest-dom": "^5.11.9",
     "@testing-library/react": "^11.2.5",
     "@testing-library/user-event": "^12.8.3",
-    "axios": "^1.9.0",
+    "axios": "^0.21.1",
     "bootstrap": "^5.1.3",
-    "gh-pages": "^6.3.0",
-    "react": "^18.3.1",
+    "gh-pages": "^3.1.0",
+    "react": "^19.1.0",
     "react-bootstrap": "^2.2.0",
-    "react-dom": "^18.3.1",
-    "react-pdf": "^9.2.1",
+    "react-dom": "^19.1.0",
+    "react-pdf": "^5.7.1",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "^0.0.0",
+    "react-scripts": "4.0.3",
     "swiper": "^9.4.1",
     "web-vitals": "^1.1.1"
   },
