@@ -1,12 +1,18 @@
 import React from "react";
-import ReactDOM from "react-dom";
+// import ReactDOM from "react-dom";
 import {HashRouter} from 'react-router-dom';
-import App from "./App";
+// import App from "./App";
+import { createRoot } from 'react-dom/client';
+import App from './App';
 
-ReactDOM.render((
+const root = createRoot(document.getElementById('root'));
+root.render(<App />);
+
+root.render((
   <React.StrictMode>
    <HashRouter basename={process.env.PUBLIC_URL}>
      <App />
    </HashRouter>
    </React.StrictMode>
 ), document.getElementById("root"));
+
