@@ -57,9 +57,9 @@ export default function Home() {
               </Col>
             </Row>
             <Row>
-              {/* <Col className="displayFlex justifyContentCenter " xs={12} xl={4}>              
+              <Col className="displayFlex justifyContentCenter " xs={12} xl={4}>              
                 <Button id="btnResume"><Link style={{textDecoration:"none"}} to="/resume">View My Resume</Link></Button>
-              </Col> */}
+              </Col>
               <Col className="displayFlex justifyContentCenter " xs={12} xl={4}>              
                 <Button id="btnResume"><Link style={{textDecoration:"none"}} to="/reviews">View My Reviews</Link></Button>
               </Col>
